remove deleted and edited menuImg from cloudinary

+++ src/pages/api/backoffice/menus/delete.ts
@@ -2,6 +2,8 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "../../auth/[...nextauth]";
 import prisma from "@/config/client";
+import cloudinary from "@/config/cloudinary";
+
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
@@ -12,14 +14,25 @@ export default async function handler(
     //if session exists
 
     const id = Number(req.query.id);
-    console.log({ menuId: id });
 
     if (!id) {
       return res.status(400).end();
     }
 
-    //delete menus and menu categories
+    //getMenuImage
+    const menuToDel = await prisma.menus.findUnique({
+      where: { id },
+    });
+    const imgToDel = menuToDel?.asset_url || "";
+    let publicId = imgToDel.split("nextjs-uploads/").pop()?.split(".")[0];
+    publicId = "nextjs-uploads/" + publicId;
+    const defaultImg = "nextjs-uploads/jyumli1sfg9t4b4qcyck";
+    if (publicId && defaultImg !== publicId) {
+      const resCloudinary = await cloudinary.uploader.destroy(publicId);
+      if (resCloudinary.result !== "ok") return res.status(400).end();
+    }
 
+    //delete menus and menu categories
     await prisma.menus_menu_categories_locations.deleteMany({
       where: { menus_id: id },
     });
@@ -30,7 +43,6 @@ export default async function handler(
       data: { is_archived: true },
     });
     if (!delMenu.id) return res.status(500).end();
-
     res.status(200).end();
   } else {
     res.status(401).end();
